workflow: add retention period
In this commit, watchers learn a new configuration attribute `retentionPeriod`
passed to the workflow as part of the blob event struct.

In `enduro.toml`, set up the retention period as follows:

```
[[watchers.foobar]]
retentionPeriod = "24h"
```

The format follows Go's convention to express duration:
https://golang.org/pkg/time/#ParseDuration.

The zero-value, e.g. when the retention period config is missing or zero, will
cause the original to not be deleted.

The implementation is based on Cadence workflow timers. The timer is launched
as long as the package processing has completed successfully. The deletion
occurs when the timer future is released.