watcher: support filesystem polling
In this pull request, filesystemWatcher learns to monitor directories using
polling when the inotify API can't be set up. It can also be enforced via
configuration, in `enduro.toml`:

```toml
[[watcher.filesystem]]
inotify = false
```

The new `filenotify` package is a copy from the one found in the Moby project,
since they haven't published it as a module. This package provides an interface
with an API similar to the fsnotify struct so the changes required were pretty
small. Credits to Brian Goff and the Moby team.