workflow: do not retry AM poll activities
This commit updates the poller activities so they convert errors that escape
the backoff retry into workflow NRE (non-retryable) errors. When these are
returned, they escape the retry policy attached to the activities.

E.g., we want the activity to give up right away when Archivematica returns the
"FAILED" status for a package.