policies: increase StartToCloseTimeout
In the case of large files the download activity can timeout because it
is not finished within 600 seconds. Increasing this timeout should
mitigate the problem of timeouts for large files.

We have also added the StartToCloseTimeout as a non retryable error.
Generally speaking if the download is not finished within the allowed
timeframe it will not finish when retrying.

Increasing the timeout is just a temporary workaround. The download
acticity should be implemented as a heartbeat activity to check for
actual progress in the downloading, not timeout based on a "wors-case"
kind of scenario.