status: add case for net.Error
The timeout is a special case of a net error. Currently we only retry on
the timout explicetly. But the net error seems to have a nice
isTemporary() function. From the go docs:

// An Error represents a network error.
type Error interface {
	error
	Timeout() bool   // Is the error a timeout?
	Temporary() bool // Is the error temporary?
}

So we might want to have both options, or one of them (don't know the
exact behaviour). Also we might want to clean the error function a bit
by using a switch on the type (if it grows further):

https://stackoverflow.com/questions/22761562/portable-way-to-detect-different-kinds-of-network-error-in-golang