pipeline: use non-blocking acquire in semaphore
Our dumb local semaphore has been showing a deadlock in staging. I can't
reproduce locally neither understand the reason. In this commit, I change the
activity so it uses the non-blocking version of the acquire method. As a result,
the activity will be executed more frequently (every two seconds). I can't see
how that is going to help but I'm running out of ideas.

We still aim to deal with proper resource usage management using monitoring
workflows but we haven't gotten to that just yet.