transfer: implement retry
With the Archivematica Transfer throttling it is possible that the
package endpoint returns a failure because there are no more transfer
slots available. By implementing the transfer activity as a retrying
activity with a heartbeat it is possible to stage too many transfers and
Enduro will keep retrying untill the transfer is send.