Force semaphore weight to be unsigned
This commit ensures that the weight of the semaphore is never below zero
which was happening when releasing more than was held. This was seen
when workflows were resumed after the process was killed and we release
when the hold had been lost.

Connects to https://github.com/artefactual-labs/enduro/issues/272.