Use disconnected ctx to release pipeline
This commit uses a disconnected context to release the pipeline. It
ensures that we're releasing even during cancellation of workflow. Under
these circumstances, the root context is cancelled.

WaitForCancellation is set to false, this prevents the activities to
fail with `TimeoutType: HEARTBEAT` when Enduro is restarted.

Connects to https://github.com/artefactual-labs/enduro/issues/272.