Allow watchers to refer to multiple pipelines
This commit allows the association of a watcher with multiple pipelines.
The processing workflow will pick one of the given pipelines randomly.

In the following example, the user provides three pipelines:

```toml
[[watcher.filesystem]]
pipeline = ["am1", "am2", "am3"]
```

For backward-compatibility, a single string is still supported:

```toml
[[watcher.filesystem]]
pipeline = "am1"
```

If `pipeline` is undefined, the workflow will pick one from the list of all
pipelines configured in the application.

When starting a new batch, the pipeline field is now optional.